requirements.txt
pymongo~=3.11.1
mysql-connector-python~=8.0.22
requests~=2.27.1
python-dotenv~=0.19.2
jsonpickle~=2.1.0
ckan~=2.9.5
setuptools~=44.1.0
